Update transformation to handle releasever better
If I want to handle this perfectly, I need to generate metadata about the yum variables during the time i reposync, and then read those values later. This is a quick fix to an issue that I might just keep for my needs, but can potentially be buggy for other users.